requirements.txt
Django==5.2.9
django-environ==0.12.0

# Static files in production
whitenoise>=6.6,<7

# Render/production WSGI server (not installed on Windows)
gunicorn>=21.2,<23; platform_system != "Windows"
